Remove some unused variables from run_attention_weights_experiment()
I accidentally re-ran the experiment, which changed the results a
bit. For one attention cell, it didn't make a difference. For the
other, the experiment didn't find any strings that produced high
attention values. This is a problem. If I had not seen that cell
producing high attention values previously I may have concluded that
it's not doing anything of interest. I've left the old results
for that cell but in the next commit, I'll change the experiment
to run over all strings in the validation set and thus make the
results more comprehensive and repeatable.