Add .gitignore for s_len256 files
We don't want these in git for now because they are too big (git lfs
can handle them but they're above the max file size for github and
the total size is above my quota). The branch s_len_256 has a commit
which includes code to split these files into 2G chunks and add them
to git-lfs. But that still runs into quota problems, so I'm not
pushing that commit.