Initial port of BlockInternalsExperiment into the new framework
It works to the point of the unit test running. This also includes
the bare minimum of port over of helper functions from the old
notebook. They've been put into classes that encapsulate the model
and the tokenizer, and the methods have been given better names. I
think this is the right start to build on and I'll port other functions
over as I need them.