Add CLI runner for block internals experiment
Added a function that implements the CLI interface to running
the block internals experiment. It's exposed as a console script.

I added a depenency on [Click](https://click.palletsprojects.com/en/8.1.x/),
which is surprisingly nice to use.

I ran the experiment over all the blocks, and had it generate files
in the same output directory as the old notebook wrote to:

```bash
for b in {0..5}; do block_internals_exp_run nbs/artifacts/shakespeare.pt nbs/artifacts/input.txt nbs/artifacts/block_internals_results/ --block_idx $b; done
```
Git shows none of the data files changed, which is a very good sign.

I also ran the analysis in _old/_transformer-analysis.ipynb that consumes
the output files and none of the output changed. So I think this
experiment is running well.