Move module in which DataaBatcher lives
Previously, I'd created a common/ directory in which I hoped
I could add a bunch of separate notebooks that all contribute
code to a "common" module. Sadly, this doesn't work. nbdev
doesn't support multiple notebooks exporting to a single
target module.

So I'm going to keep the comoon directory and have all the
notebooks in it export to submodules off common. As a start,
I'm moving DataBatcher into a dedicated "databatcher"
submodule.