Get rid of EncodingHelpers.embed_string()
I realized that the embed methods should be on TransfomerAccessors,
but I can't support embed_string() there because I'd like to keep
TransformerAccessors independent of the tokenizer. embed_string()
is just a tiny wrapper around embed_tokens(tokenize_string(s)) so
not a big loss. Just get rid of it.