Big rewrite of BlockInternalsExperiment and clustering writeup
A few findings that motivated this:
- After analyzing the clustering results I had for several days, no
  good patterns had emerged. Realized that instead of clustering, I
  can just directly identify the proj output/ffwd values for other
  tokens that are closest to that of the prompt I'm analyzing.
- After doing that, I realized I didn't have enough data (was
  previously going over the validation set with stride 96).Reduced
  that, but realized a bigger problem is that I'm only looking at
  the validation set, which is just the last 10% of the text. Because
  the text is all of Shakespeare's plays, it's not evenly distributed
  i.e. there are names, words etc, in the earlier plays that I
  wouldn't see if just looking at the validation set.
- I realized that the whole BlockInternalsExperiment can run against
  all the length 3 substrings in like 3 seconds, and easily fits in
  memory, so the whole mechanism of writing files to disk etc is
  totally unnecessary.

So BlockInternals experiment now just takes a set of strings and
runs them all in one batch. And I rewrote the writeup to do the
distance to target thing instead of clustering. More analysis
coming up.

This also includes the ability to plot subsets of a
BlockInternalsAnalysis because I needed that during some investigations
but is not actually used in the real writeup yet.

Though BlockInternalsExperiment now doesn't write/read files, I'm
leaving the data files in artifacts/block_internals_results in
place because _old/_transformer_analysis.ipynb uses them. They're
small anyway, so should be fine.