Create BatchedBlockInternalsExperiment to run with longer prompt lengths
I tried to run the existing in-memory experiment with string
length 7 and it crashed the Jupyter kernel.

Was able to run in-memory with length 5 and saw really promising
results, so decided it was worth trying for longer.

Wrote BatchedBlockInternalsExperiment which writes the results
to disk and is able to do the similarity analysis without
having to read everything into memory. This took a long time to
get right but I'm happy with where it ended.

I ran it on string length 10. It took about 23 minutes to run.
The resulting data files are >300 GB, so I won't be putting them
in source control. Instead I have added a makefile rule to make
it easy to regenerate.

Made some changes to simulate() to allow using prompts that are
not in the set of strings in the experiment. Also had to write
a simulate2() to work with the batched experiments. Unfortunately,
it is slow.

The whole writeup now has to be skipped during testing because it
relies on these large datasets that I'm not checking in.