Make topk_across_batches work when the results have arbitrary dimensions
It previously assumed that the results from process_batch() are a
1-D tensor. Now it can be any shape, as long as the first dimension
is the same as the batch size.

This is a precursor to writing a version of simulate() that can run
multiple strings at the same time. We'd be evaluating
BatchedBlockInternalsExperiment.strings_with_topk_closest_*()
with queries that have values from multiple strings, so that we
can perform multiple distance calculations with each batch loaded
into memory.