Wrote simulate3()
This version takes multiple prompts and runs the simulation and model
on them all at the same time. This should be much faster, because when
we load a batch from disk, we run the analysis for all prompts against
it at the same time. The older simulate2() would have loaded the batch
from disk once for each prompt.

Tested by comparing output of simulate3() on a bunch of prompts to the
output of simulate2() run on those prompts individually.