Add manim as a dep and hanges to the requirements files
I tried to add manim to the dependencies
and this resulted in a bunch of conflicts
that required changes.

1. Added `--resolver=backtracking` to the
  `pip-compile` commands

   This is apparently the new resolver. I'd
   been using the legacy one and ignoring
   the warnings about it being deprecated.
   This solved some of the problems.

2. Upgraded the version of pip-tools

   I tried to do pip-sync and it failed
   without this. After doing this, I
   `pip-compiled` both files again to get
   the latest version

3. Set requirements.dev.in to be constrained
   by the packages installed in
   requirements.txt. Without this, there were
   conflicting dependencies generated by
   compiling requirements.dev.in after
   requirements.in.

4. Finally added manim as a dependency and
   compiled everything again.

5. After all these changes, ran:

```
pip-sync requirements.txt requirements.dev.txt
```

   to synchronize the virtual environment
   with what the requirements files say.

A bunch of package versions got updated
as a result of this. I sanity checked a few
of the older notebooks and it seems like
things work but sadly some of the image data
has changed (matplotlib upgraded). Will try
to re-run all these and commit them at once
in a separate commit.