A mostly failed approach to generating similar strings
Most of this is going to end up getting undone, but I'm
committing it so it lives in history as there's a lot of
code I may want to refer to.

The key things motivated this work: last week I realized
I had to compute similar strings for all the t_i indexes
for the proj_out and ffwd_out values.

My first approach to this was to just run a block internals
experiment for each substring length. This ended up using
all my disk space, so I started searching for ways to do it
with the data from just one block internals experiment,
ideally without having to generate slices. The last several
commits include some perf measurments that showed I might
be able to load all the proj_out / ffwd_out data as memory
mapped tensors and then evaluate all strings against this
loaded data and get some perf wins.

The approach in this commit is an implementation of that.
Turns out its terrible. While my perf tests ran will with
4 strings and were faster than the other methods on 4 strings,
what I failed to check was how it would do on a batch of 100
strings. It doesn't work at all - just crashed last night
even before running a single one. (The embeddings similar
strings ran, but that doesn't use this new approach - its
just too much data. That should have been a clue).

This led to some work today to separate the generation steps
into separate CLI commands that can run:
- potentially at the same time
- free of any other cruft in memory that a Jupyter notebook
  might have

I was optimistic this would run. This crashed too, right
away. You just can't run a query that big on a 64GB RAM
laptop.

Turning the batch size down to 10 does run, but it's not
_that much_ faster. I ran the proj_out query with batch
size 10 and it took 5:04:24 for 20,000 strings. So about
0.9 seconds per string.

Going way back, all of this optimization effort was about
trying to make simulate3() fast enough to compute these
similar strings interactively. I think I've realized that
is the wrong approach. I should pre-compute them in batch
for a bunch of strings and then simulate_n() would just
work on those pre-computed values nice and fast. So in that
world, whether it takes .9 seconds or 2 seconds doesn't
make a whole lot of differennce, because it's a one-time
cost.

The only comparison I have is that when I last run 20,000
strings through simulate3(), it took about 2.5 minutes per
batch of 100 strings, to do the embeddings, proj_out, and
ffwd_out for each. That was using the slicer.

With the mindset that it's a fixed, one-time cost, it's
not even clear I need the slicer. I'd like to get rid of
it for simplicity and to save disk space.

I'm going to investigate this next. But I'm committing
this for history and because I bet there is some useful
code here.