Go back to non-slicer method for proj/ffwd output similars
Per the previous commit message, the optimizations I tried
didn't work. In this commit, I've added back the methods on
BatchedBlockInternalsExperiment to do the topk closest string
calculations based on the unsliced batches. I wired these up
to SimilarStringsExperiment and the cli. Ran it overnight.

The proj/ffwd out generation to 5 hours and 43 minutes but
ran concurrently and didn't put the laptop under memory
pressure.

Things are now in a bit of a mess because we have 3 ways of
doing these calculations:
- The way added in this commit
- The way that mmap's the all batches into memory (previous
  commit)
- The slicer

I'll be cleaning these up after I convert all code that uses
the slicer to use the precomputed results.