Replace simulate3() with version that runs over SimilarStringExperiment results
I've implemented a version of simulate3 that uses takes
SimilarStringsExperiment results, computes frequencies
and returns the simulated and model outputs. I've updated
the clustering analysis notebook to use this, verifying
that the output hasn't changed for known strings. To
make this work, I had to run a new, small
SimilarStringsExperiment over the set of strings used in
the notebook.

The stuff around analyzing the result data files from a
previous informal experiment to run simulate3() and save
the intermediates in json files is still there but outdated.
I'll remove it as part of a bigger cleanup.