Make topk_across_batches work with batches of different sizes
In the case when we call strings_with_topk_closest_proj_outputs()
or strings_with_topk_closest_ffwd_outputs() with a t_i argument
that isn't -1, we will want to only evaluate the unique substrings
in each batch. This means that each batch will have a different
number of items. This change is in prep for that.