Big refactor of the simulation code
This separates out the building of the frequency tensors
based on the similar strings from the logic that combines
those frequencies to produce the final next token probability
distribution. This makes it possible to easily try out
different variations of logic. In this commit, I've
implemented one that uses inverse distances to weight
the frequency tensors. It doesn't work as well, but it's
an example of the framework in use.

This also runs fast now so I don't need the cells I'd
written to run the simulation on a sample.