Change existing code to use new dataset
Yesterday I kicked off a new run that generates
the ffwd outputs for more data. The first chunk
of these is the same as the ffwd data I'd generated
before. This commit switches the existing analysis
to use this.

Because I'm no longer generating emb or proj out
data, I've removed the analysis that refers to
them from the notebook. They weren't useful in
predicting output.