Add training code to transformer notebook
This code was actually run in a PaperSpace GPU
environment to train the model. In PaperSpace,
I just copied all the other imported stuff
from the project into the notebook so I didn't
have to clone the whole repo. Output shown is
from the actual training run.

This also includes a change to the
TransformerLanguageModel class to better initialize
the weights. This was based on this note I saw in
Andrej Karpathy's repo:
https://github.com/karpathy/ng-video-lecture#nanogpt-lecture