Add exported CosineSimilaritiesExperiment and cli
As opposed to the janky one I was running from inside
40_widening_similar_space.ipynb, this one is exported
as a module and has a real CLI. It is also written to
be independent of block internals experiment. i.e. it
generates its own data rather than using data from a
block internals experiment. This makes it easier to
run on remote hosts without transferring gigabytes of
block internals data. The downside is it seems to be
a lot slower. Next, will try to make it run on the
GPU and see if it goes faster.