Make nbdev_test pass on GPU machine
Because I've always run everything on my laptop
which does not have a cuda device, everything
has always been on the CPU. I'd gotten sloppy
about adding the appropritate `device=` and
`.cpu()` calls to move things to the correct
devices. This commit fixes this and makes it
so that nbdev_test passes on a machine with both
cuda and cpu devices.