Use correct model
I realized that all the analysis from before
vacation was comparing to model outputs from
the wrong model!

This updates the model paths and re-runs some
of the data with the same thresholds. It's
clear these thresholds need adjusting but it's
too slow and painful with the current approach.
Going to work on making that faster and then
will re-run everything.