Implement prefiltering of cosine sim results
This commit contains the code to implement
prefiltering and the changes to 40_widening_similar_space.ipynb
to use it.

As of this commit, I've just changed the code
to filter over the prefiltered results but have
not changed any of the thresholds. The results
are the unchanged, which is a good indication
that prefiltering is working correctly. Now,
the analyze_dataset() call takes 3 seconds
instead of ~80 seconds, so I can tune interactively.