Move large generated files out of nbs/ directory
Turns out that nbdev_preview was copying the whole
of the nbs/ directory to _proc/ and with all the
large files there, that was taking forever and
consuming a ton of disk space:

(venv-transformer) (base) [spather@Shyamalans-MacBook-Pro-2 (master) ~/local-storage/code/transformer-experiments/nbs/artifacts]
$ gdu -h --max-depth=1
77M     ./logits_results
2.1G    ./model-training
385G    ./cosine_sim_results
110G    ./attn_weight_results
385G    ./block_internals_results
881G    .

The fix is to create a generated_data/ directory
in the root, gitignore everything in it, and
move all the large files into this directory. With
the prior changes to move everything to use the
environment API, the change is relatively simple
because we (mostly) just have to change the value
of data_root for the relevant environments.

Note that for any artifacts that were checked in,
I'm leaving these in nbs/artifacts. This is mostly
old stuff and it's small, so not a big deal. To
accommodate these being left behind, I had to change
the code that touches them to use

`environment.code_root / 'nbs/artifacts/...'`

rather than data_root.

I also updated the makefile so that these files
will get generated in the new location if run
again.

Since the whole of the new directory is covered
by the root .gitignore, I am removing a bunch of
other .gitignore files that covered stuff within
nbs/.

After this change, this is the disk usage of
nbs/artifacts:

$ gdu -h --max-depth=1
77M     ./logits_results
94M     ./block_internals_results
322M    .

And nbdev_preview now runs again!