Pre-filter all queries
Previous, I'd only run the pre-filtering for the first
500 queries. To compute `ffwd_freqs` I ran it over the
whole cosine sims experiment result set. In cleaning
this up for the blog, I wanted to be able to run everything
over prefiltered data for speed.

This change does two things:
- Adds code to do prefiltering for all the queries (takes
  about 10 hours to run)
- Changes the calculation of ffwd_freqs to use the
  new prefiltered data

Re-ran a bunch of the analyses after computing ffwd_freqs
this way and nothing changed.