Update a bunch of analysis with better hand-rolled coefficients
After determining the best-handrolled coefficients on
each block separately on a sample of 500 queries, I
ran the simulation against all 20k queries. I got results
roughly in line with what I got on the sample of 500.

But then, I further tweaked the coefficients for the 20k
data set and was able to get slightly better results.
Turns out, in a bunch of the later analysis, I was not
using these better tweaked coefficients. This change fixes
that and as a result some of the specific results changed.