Move model instantiation to top of blog post
The drawing code needs the model and tokenizer. Since
I moved the drawing code up, I need the model instantiated
higher up. This isn't great - makes for a worse flow in
the text. But I'm just moving this for now to get things
working and will address it in the edit pass.

Probably the right thing to do is have the drawing code
instantiate its own model and tokenizer so that it
will be independent of where we instantiate the model for
the rest of the notebook's use.