Include example where the approximation using the first-layer only doesn't work well
When writing up the initial approximation using only the
first layer, I picked an example that did particularly
well. The problem is that when including all layers, this
example doesn't improve and actually gets worse. I know
this is not true for all examples, so I wrote some code
to find examples where there's a bad approximation at
layer 0 but ultimately a good one when considering all
layers.

In this commit, I included the code that finds the examples.
In the next commit, I'll remove that code because it
isn't part of the narrative but I wanted to include it
so it's in the history.