Utilities for translating tokens into valid filenames
This is useful for situations in which we want to create
files that correspond to tokens. It deals with special
characters that can't be in filenames and distinguishes
upper and lowerspace characters so that we don't get
clashes on case-insensitve filesystems.

I wasn't sure where this should live, but since it's
so specific to the TinyShakespeare data set, I put it in
`trained_models.tinyshakespeare_transformer`.