Hot mess of WIP analysis
This is a whole bunch of work from last week on trying to learn
embeddings at different stages of the pipeline and using them
to approximate the embedding space for different tokens. I'm about
to clean all of this up but am committing it in this state first
so I don't have to feel bad about deleting things.