Export training hyperparams and estimate_loss() from transformer notebook
This makes it possible to use these for training in other notebooks.
These are exported to a new module (`models.transformer_training`), but
still live in the main transformer.ipynb notebook (mainly because I
don't want to do the work of moving them). Nbdev requires every module
to have a dedicated notebook, hence the new `tranformer-training.ipynb`
which just contains the required imports.