Fix bug in aggregate performance analysis for approximation interpretation
We want to check that the top tokens predicted by similar
strings are in the top half of the token subspaces. But we
weren't actually returning that number of token subspaces
to compare against, which depressed the results.

Also changed top_n to 5 instead of 3. I think this is a
better comparison and with the bug above fixed, results
are still good.