Improve code that compute aggregate performance of the interpreation hypothesis
Previously, the code that computed the metric for each block was
copy-pasted and each copy overwrote local variables created by
the earlier ones. This made it difficult to play with earlier
results after computing later ones. To address this, I put all
the logic into functions and created cells that create distinct
local data for each block.

In doing this, I found a major bug: I was using the block 5
projection matrices for blocks 3 and 4. Oops. This is why we
shouldn't copy paste.

After all this refactoring, the results for block 5 are the same,
which gives me confidence the refactor is correct. Blocks 3 and 4
regressed, but I believe this is a more accurate view of their
performance.