Fix kde plots so that lines for different series show up in different colors
TIL: If you're trying to plot two series on the same KDE plot by
doing:

```
    sns.kde_plot(series1, label='series 1')
    sns.kde_plot(series2, label='series 2')
```

The right thing will happen if series1 and series2
have shape (N,) - you'll get two lines in different colors.
If the shapes of series1 and series2 are (N,1), you'll get
lines of the same color.